Fix config to appConfig
