+++ client/src/api.ts
@@ -5,7 +5,7 @@ Permission to use, copy, modify, and distribute this software and its documentat
 The full terms of this copyright and license should always be found in the root directory of this software deliverable as "license.txt" and if these terms are not found with this software, please contact the USC Stevens Center for the full license.
 */
 import axios, { AxiosResponse } from "axios";
-import { Config } from "config";
+import { AppConfig } from "config";
 import {
   DeleteLesson,
   DeleteSession,
@@ -47,22 +47,25 @@ function findOrThrow<T>(res: AxiosResponse<GQLResponse<T>>): T {
   return res.data.data;
 }
 
-export async function fetchConfig(): Promise<Config> {
-  const gqlRes = await axios.post<GQLResponse<{ config: Config }>>("", {
-    query: `
+export async function fetchAppConfig(): Promise<AppConfig> {
+  const gqlRes = await axios.post<GQLResponse<{ appConfig: AppConfig }>>(
+    GRAPHQL_ENDPOINT,
+    {
+      query: `
       query {
-        config {
+        appConfig {
           googleClientId
         }
       }
     `,
-  });
+    }
+  );
   if (gqlRes.status !== 200) {
-    throw new Error(`config load failed: ${gqlRes.statusText}}`);
+    throw new Error(`appConfig load failed: ${gqlRes.statusText}}`);
   }
   if (gqlRes.data.errors) {
     throw new Error(
-      `errors reponse to config query: ${JSON.stringify(gqlRes.data.errors)}`
+      `errors reponse to appConfig query: ${JSON.stringify(gqlRes.data.errors)}`
     );
   }
   if (!gqlRes.data.data) {
@@ -70,7 +73,7 @@ export async function fetchConfig(): Promise<Config> {
       `no data in non-error reponse: ${JSON.stringify(gqlRes.data)}`
     );
   }
-  return gqlRes.data.data.config;
+  return gqlRes.data.data.appConfig;
 }
 
 export async function fetchSessions(
