+++ cypress/cypress/fixtures/users.ts
@@ -4,20 +4,39 @@ Permission to use, copy, modify, and distribute this software and its documentat
 
 The full terms of this copyright and license should always be found in the root directory of this software deliverable as "license.txt" and if these terms are not found with this software, please contact the USC Stevens Center for the full license.
 */
-import { fetchAppConfig } from "api";
 
-export interface AppConfig {
-  googleClientId: string;
-}
-
-export async function getClientID(): Promise<string> {
-  if (process.env.GOOGLE_CLIENT_ID) {
-    return process.env.GOOGLE_CLIENT_ID;
-  }
-  try {
-    return (await fetchAppConfig()).googleClientId;
-  } catch (err) {
-    console.error(err);
-    throw err;
-  }
+export const users = {
+  edges: [
+    {
+      cursor: "cursor 1",
+      node: {
+        id: "admin",
+        name: "Admin",
+        email: "admin@opentutor.org",
+        userRole: "admin",
+      },
+    },
+    {
+      cursor: "cursor 2",
+      node: {
+        id: "contentmanager",
+        name: "Content Manager",
+        email: "contentmanager@opentutor.org",
+        userRole: "contentManager"
+      },
+    },
+    {
+      cursor: "cursor 2",
+      node: {
+        id: "author",
+        name: "Author",
+        email: "author@opentutor.org",
+        userRole: "author"
+      },
+    },
+  ],
+  pageInfo: {
+    hasNextPage: false,
+    endCursor: "cursor 2",
+  },
 }
