+++ cypress/cypress/integration/login.spec.ts
@@ -5,7 +5,7 @@ Permission to use, copy, modify, and distribute this software and its documentat
 The full terms of this copyright and license should always be found in the root directory of this software deliverable as "license.txt" and if these terms are not found with this software, please contact the USC Stevens Center for the full license.
 */
 import { lessons } from "../fixtures/lesson";
-import { cySetup, cyMockDefault, mockGQL } from "../support/functions";
+import { cySetup, cyMockDefault, mockGQL, mockGQLConfig, CONFIG_DEFAULT } from "../support/functions";
 
 describe("Login", () => {
   it("loads home page", () => {
@@ -27,7 +27,6 @@ describe("Login", () => {
       noLogin: true
     })
     cy.visit("/");
-    cy.intercept("**/config", { GOOGLE_CLIENT_ID: "test" });
     cy.get("#login-menu #login-button").should("not.be.disabled");
   });
